refactor: Move Room database components to dedicated files
This commit refactors the Room database setup by moving `AppDatabase` and `AnimeDao` to their own files within the `com.branwen.mal.models` and `com.branwen.mal.interfaces` packages, respectively.

Key changes:
- `AppDatabase.kt`: New file created in `com.branwen.mal.models` to house the `AppDatabase` class definition.
- `AnimeDao.kt`: New file created in `com.branwen.mal.interfaces` to house the `AnimeDao` interface.
- `AnimeRepository.kt`: Removed the inline definitions of `AnimeListEntity`, `AnimeDao`, and `AppDatabase`. It now relies on the new, separate files.
- `AnimeLocalDataSource.kt`: Updated import paths to reflect the new location of `AnimeDao` and `AnimeListEntity`.
- `AppModule.kt`: Updated import paths for `AnimeDao` and `AppDatabase`.
- `.gitignore`: Added `.kotlin/` to ignore Kotlin build-related files.

This restructuring improves code organization and modularity by placing database-related components in their designated packages.
