+++ app/src/main/java/com/branwen/mal/data/repo/AnimeRepository.kt
@@ -1,13 +1,5 @@
 package com.branwen.mal.data.repo
 
-import androidx.room.Dao
-import androidx.room.Database
-import androidx.room.Entity
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.PrimaryKey
-import androidx.room.Query
-import androidx.room.RoomDatabase
 import com.branwen.mal.data.local.AnimeLocalDataSource
 import com.branwen.mal.data.remote.AnimeRemoteDataSource
 import com.branwen.mal.models.AnimeNode
@@ -17,6 +9,16 @@ import kotlinx.coroutines.flow.emitAll
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.flow
 
+/**
+ * Repository for handling anime data, acting as a single source of truth.
+ *
+ * This class abstracts the data sources (remote and local) and provides a clean API
+ * for accessing and managing anime information. It handles fetching data from the remote
+ * source, caching it locally, and retrieving it from the local cache when appropriate.
+ *
+ * @param remote The remote data source for fetching anime data from an external API.
+ * @param local The local data source for storing and retrieving cached anime data.
+ */
 class AnimeRepository(
     private val remote: AnimeRemoteDataSource,
     private val local: AnimeLocalDataSource
@@ -45,48 +47,4 @@ class AnimeRepository(
             local.getAnimeDetails(animeId) ?: error("No details cached")
         }
     }
-}
-
-@Entity(tableName = "anime_list")
-data class AnimeListEntity(
-    @PrimaryKey val id: Int,
-    val title: String,
-    val status: String,
-    val imageUrl: String,
-    val startSeason: String,
-    val startYear: String,
-    val numEpisodesWatched: Int,
-    val totalEpisodes: Int?,
-    val rating: Int
-)
-
-@Dao
-interface AnimeDao {
-    @Query(
-        value = "SELECT * FROM anime_list " +
-                "ORDER BY LOWER(TITLE) ASC, " +
-                "CASE status " +
-                "   WHEN 'watching' THEN 1 " +
-                "   WHEN 'completed' THEN 2 " +
-                "   WHEN 'on_hold' THEN 3 " +
-                "   WHEN 'dropped' THEN 4 " +
-                "   ELSE 5 " +
-                "END"
-    )
-    fun getAll(): Flow<List<AnimeListEntity>>
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertAll(list: List<AnimeListEntity>)
-
-    @Query("DELETE FROM anime_list")
-    suspend fun clearAll()
-}
-
-@Database(
-    entities = [AnimeListEntity::class],
-    version = 1,
-    exportSchema = false
-)
-abstract class AppDatabase : RoomDatabase() {
-    abstract fun animeDao(): AnimeDao
 }
